# admin-backend/docs/swagger.yaml
definitions:
...
  admin_proto.AdminInfo:
    properties:
      adminId:
        type: integer
      avatar:
        type: string
      createdAt:
        type: string
      email:
        type: string
      enabled:
        type: boolean
      expire:
        type: integer
      expireDataTime:
        type: string
      lastLoginIp:
        type: string
      lastLoginTime:
        type: string
      loginTotal:
        type: integer
      menus:
        additionalProperties:
          $ref: '#/definitions/admin_proto.MenuItem'
        type: object
      nickname:
        type: string
      permissions:
        additionalProperties:
          type: string
        type: object
      roles:
        items:
          $ref: '#/definitions/admin_proto.AdminUserRoleItem'
        type: array
      token:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
  admin_proto.AdminUserListItem:
    properties:
      adminId:
        description: 管理员ID
        type: integer
      avatar:
        description: 头像
        type: string
      createdAt:
        description: 创建时间
        type: string
      email:
        description: 邮箱
        type: string
      isEnabled:
        description: 是否启用
        type: boolean
      isEnabledButtonDisabled:
        type: boolean
      lastLoginIp:
        description: 上次登录IP
        type: string
      lastLoginTime:
        description: 上次登录时间
        type: string
      loginTotal:
        description: 登录次数
        type: integer
      nickname:
        description: 昵称
        type: string
      roles:
        items:
          $ref: '#/definitions/admin_proto.AdminUserRoleItem'
        type: array
      updatedAt:
        description: 更新时间
        type: string
      username:
        description: 账号名称
        type: string
    type: object
  admin_proto.AdminUserRoleItem:
    properties:
      roleId:
        description: 角色ID
        type: integer
      roleName:
        type: string
    type: object
  admin_proto.MenuItem:
    properties:
      authority:
        type: string
      component:
        type: string
      hideChildrenInMenu:
        type: boolean
      hideInMenu:
        type: boolean
      icon:
        description: 菜单图标
        type: string
      key:
        description: 菜单唯一键
        type: string
      name:
        description: 菜单名称
        type: string
      path:
        description: 菜单路由
        type: string
      routes:
        items:
          $ref: '#/definitions/admin_proto.MenuItem'
        type: array
    type: object
...
  admin_proto.RespAccountInfoData:
    properties:
      data:
        $ref: '#/definitions/admin_proto.AdminInfo'
    type: object
...
  admin_proto.RespAdminUserListData:
    properties:
      list:
        items:
          $ref: '#/definitions/admin_proto.AdminUserListItem'
        type: array
      pageNum:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  code.Message:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/code_proto.ErrorCode'
        description: 错误码
      data: {}
      msg:
        description: 错误码描述
        type: string
      reason:
        description: 错误具体信息
        type: string
    type: object
  code_proto.ErrorCode:
    enum:
    - 0
    - 1
    - 100001
    - 100002
    - 100003
    - 200001
    - 200002
    - 200003
    - 200004
    - 200005
    - 300001
    - 300002
    - 300003
    - 400001
    - 400002
    - 400003
    - 500001
    - 500002
    - 500003
    - 500004
    - 600001
    - 600002
    - 600003
    - 600004
    - 600005
    - 700001
    - 800001
    type: integer
    x-enum-comments:
      ErrorCode_AdminAccountInvalid: 账号已被封禁
      ErrorCode_AdminAccountNotExist: 账号不存在或已被删除
      ErrorCode_AdminAccountPasswordInvalid: 账号或密码错误
      ErrorCode_AdminApiKeyExist: 接口键名已存在
      ErrorCode_AdminApiNameExist: 接口名称已存在
      ErrorCode_AdminApiNotExist: 接口不存在货已失效
      ErrorCode_AdminApiPathExist: 接口路径已存在
      ErrorCode_AdminMenuNotExist: 菜单不存在
      ErrorCode_AdminPermissionExist: 该菜单下权限已存在
      ErrorCode_AdminPermissionKeyExist: 权限键名已存在
      ErrorCode_AdminPermissionKeyNeed: 权限键名不能为空
      ErrorCode_AdminPermissionNameNeed: 权限名称不能为空
      ErrorCode_AdminPermissionTypeInvalid: 权限类型无效
      ErrorCode_AdminRoleNotExist: 角色不存在
      ErrorCode_AuthTokenFailed: 鉴权失败
      ErrorCode_AuthTokenForbidden: 没有访问权限
      ErrorCode_AuthTokenInfoInvalid: Token信息无效
      ErrorCode_AuthTokenInspectInvalid: BearerToken无效
      ErrorCode_AuthTokenInvalid: BearerToken无效
      ErrorCode_Error: 失败
      ErrorCode_ReadContextRequestBodyFailed: 获取请求体内容失败
      ErrorCode_RecordNValidCanNotDeleted: 生效中的记录不能删除
      ErrorCode_RecordNotExist: 记录不存在
      ErrorCode_RequestBodyInvalid: 请求体参数无效
      ErrorCode_RequestParamsInvalid: 请求参数无效
      ErrorCode_RequestQueryInvalid: 请求查询参数无效
      ErrorCode_Success: 成功
    x-enum-varnames:
    - ErrorCode_Success
    - ErrorCode_Error
    - ErrorCode_ReadContextRequestBodyFailed
    - ErrorCode_RecordNotExist
    - ErrorCode_RecordNValidCanNotDeleted
    - ErrorCode_AuthTokenFailed
    - ErrorCode_AuthTokenInvalid
    - ErrorCode_AuthTokenInspectInvalid
    - ErrorCode_AuthTokenInfoInvalid
    - ErrorCode_AuthTokenForbidden
    - ErrorCode_RequestBodyInvalid
    - ErrorCode_RequestQueryInvalid
    - ErrorCode_RequestParamsInvalid
    - ErrorCode_AdminAccountPasswordInvalid
    - ErrorCode_AdminAccountNotExist
    - ErrorCode_AdminAccountInvalid
    - ErrorCode_AdminApiNameExist
    - ErrorCode_AdminApiPathExist
    - ErrorCode_AdminApiKeyExist
    - ErrorCode_AdminApiNotExist
    - ErrorCode_AdminPermissionKeyExist
    - ErrorCode_AdminPermissionExist
    - ErrorCode_AdminPermissionKeyNeed
    - ErrorCode_AdminPermissionNameNeed
    - ErrorCode_AdminPermissionTypeInvalid
    - ErrorCode_AdminMenuNotExist
    - ErrorCode_AdminRoleNotExist
info:
  contact: {}
paths:
...
  /admin/user/add:
    post:
      consumes:
      - application/json
      description: 添加管理员
      parameters:
      - description: '@gotags: label:"头像地址" binding:"omitempty,url"'
        in: query
        name: avatar
        type: string
      - description: '@gotags: label:"确认密码" binding:"required,eqfield=Password,adminpwd"'
        in: query
        name: confirmPassword
        required: true
        type: string
      - description: '@gotags: label:"邮箱" binding:"omitempty,email"'
        in: query
        name: email
        type: string
      - description: '@gotags: label:"账号启用状态"'
        in: query
        name: enabled
        type: boolean
      - description: '@gotags: label:"昵称" binding:"required,trimBlank"`'
        in: query
        name: nickname
        required: true
        type: string
      - description: '@gotags: label:"密码" binding:"required,adminpwd"`'
        in: query
        name: password
        required: true
        type: string
      - description: '@gotags: label:"账号" binding:"required,adminname"`'
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 添加管理员
      tags:
      - 管理员账号接口相关
  /admin/user/bind_roles:
    post:
      consumes:
      - application/json
      description: 管理员绑定角色
      parameters:
      - description: '管理员ID @gotags: label:"管理员ID" binding:"required,min=1"'
        in: query
        minimum: 1
        name: adminId
        required: true
        type: integer
      - collectionFormat: csv
        description: '角色ID列表 @gotags: label:"角色ID列表" binding:"required,min=1"'
        in: query
        items:
          type: integer
        minItems: 1
        name: roleIds
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 管理员绑定角色
      tags:
      - 管理员账号接口相关
  /admin/user/delete:
    post:
      consumes:
      - application/json
      description: 删除管理员
      parameters:
      - description: '管理员ID @gotags: label:"管理员ID" binding:"required,min=1"'
        in: query
        minimum: 1
        name: adminId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 删除管理员
      tags:
      - 管理员账号接口相关
  /admin/user/edit:
    post:
      consumes:
      - application/json
      description: 编辑管理员
      parameters:
      - description: '管理员ID @gotags: label:"管理员ID" binding:"required,min=1"'
        in: query
        minimum: 1
        name: adminId
        required: true
        type: integer
      - description: '头像地址 @gotags: label:"头像地址" binding:"omitempty,url"'
        in: query
        name: avatar
        type: string
      - description: '邮箱 @gotags: label:"邮箱" binding:"omitempty,email"'
        in: query
        name: email
        type: string
      - description: '账号启用状态 @gotags: label:"账号启用状态"'
        in: query
        name: enabled
        type: boolean
      - description: '昵称 @gotags: label:"昵称" binding:"required,trimBlank"'
        in: query
        name: nickname
        required: true
        type: string
      - description: '账号 @gotags: label:"账号" binding:"required,adminname"'
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 编辑管理员
      tags:
      - 管理员账号接口相关
  /admin/user/edit_pwd:
    post:
      consumes:
      - application/json
      description: 编辑管理员密码
      parameters:
      - description: '管理员ID @gotags: label:"管理员ID" binding:"required,min=1"'
        in: query
        minimum: 1
        name: adminId
        required: true
        type: integer
      - description: '确认密码 @gotags: label:"确认密码" binding:"required,eqfield=Password,adminpwd"'
        in: query
        name: confirmPassword
        required: true
        type: string
      - description: '密码 @gotags: label:"密码" binding:"required,adminpwd"'
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 编辑管理员密码
      tags:
      - 管理员账号接口相关
  /admin/user/enable:
    post:
      consumes:
      - application/json
      description: 启用|禁用管理员账号
      parameters:
      - description: '管理员ID @gotags: label:"管理员ID" binding:"required,min=1"'
        in: query
        minimum: 1
        name: adminId
        required: true
        type: integer
      - description: '账号启用状态 @gotags: label:"账号启用状态"'
        in: query
        name: enabled
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 启用|禁用管理员账号
      tags:
      - 管理员账号接口相关
  /admin/user/info:
    post:
      consumes:
      - application/json
      description: 管理员详情
      parameters:
      - description: '@gotags: label:"管理员ID" binding:"required,min=1"`'
        in: query
        minimum: 1
        name: adminId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  $ref: '#/definitions/admin_proto.RespAccountInfoData'
              type: object
      summary: 管理员详情
      tags:
      - 管理员账号接口相关
  /admin/user/list:
    post:
      consumes:
      - application/json
      description: 管理员列表
      parameters:
      - description: 邮箱
        in: query
        name: email
        type: string
      - description: 0全部，1启用，2禁用
        in: query
        name: enabled
        type: integer
      - description: 昵称
        in: query
        name: nickname
        type: string
      - collectionFormat: csv
        description: 角色ID
        in: query
        items:
          type: integer
        name: roleIds
        type: array
      - description: 账号
        in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  $ref: '#/definitions/admin_proto.RespAdminUserListData'
              type: object
      summary: 管理员列表
      tags:
      - 管理员账号接口相关
swagger: "2.0"